+++ src/models/scallopUtils.ts
@@ -2,7 +2,6 @@ import { SUI_TYPE_ARG, normalizeStructTag } from '@mysten/sui.js/utils';
 import { SuiKit } from '@scallop-io/sui-kit';
 import { SuiPriceServiceConnection } from '@pythnetwork/pyth-sui-js';
 import { ScallopAddress } from './scallopAddress';
-import { ScallopQuery } from './scallopQuery';
 import {
   ADDRESSES_ID,
   PROTOCOL_OBJECT_ID,
@@ -19,7 +18,7 @@ import {
   SUPPORT_SCOIN,
   sCoinIds,
 } from '../constants';
-import { queryObligation } from '../queries';
+import { getPythPrice, queryObligation } from '../queries';
 import {
   parseDataFromPythPriceFeed,
   isMarketCoin,
@@ -31,7 +30,6 @@ import { ScallopCache } from './scallopCache';
 import { DEFAULT_CACHE_OPTIONS } from 'src/constants/cache';
 import type {
   ScallopUtilsParams,
-  ScallopInstanceParams,
   SupportCoins,
   SupportAssetCoins,
   SupportMarketCoins,
@@ -41,6 +39,7 @@ import type {
   PriceMap,
   CoinWrappedType,
   SupportSCoin,
+  ScallopUtilsInstanceParams,
 } from '../types';
 import type { SuiAddressArg, SuiTxArg, SuiTxBlock } from '@scallop-io/sui-kit';
 
@@ -60,36 +59,42 @@ export class ScallopUtils {
   public readonly params: ScallopUtilsParams;
   public readonly isTestnet: boolean;
 
-  private _suiKit: SuiKit;
-  private _address: ScallopAddress;
-  private _query: ScallopQuery;
+  public suiKit: SuiKit;
+  public address: ScallopAddress;
+  public cache: ScallopCache;
   private _priceMap: PriceMap = new Map();
-  private _cache: ScallopCache;
 
   public constructor(
     params: ScallopUtilsParams,
-    instance?: ScallopInstanceParams
+    instance?: ScallopUtilsInstanceParams
   ) {
-    this.params = params;
-    this._suiKit = instance?.suiKit ?? new SuiKit(params);
-    this._cache =
-      instance?.cache ?? new ScallopCache(DEFAULT_CACHE_OPTIONS, this._suiKit);
-    this._address =
-      instance?.address ??
-      new ScallopAddress(
-        {
-          id: params?.addressesId || ADDRESSES_ID,
-          network: params?.networkType,
-        },
-        this._cache
-      );
-    this._query =
-      instance?.query ??
-      new ScallopQuery(params, {
-        suiKit: this._suiKit,
-        address: this._address,
-        cache: this._cache,
-      });
+    this.params = {
+      pythEndpoints: params.pythEndpoints ?? PYTH_ENDPOINTS['mainnet'],
+      ...params,
+    };
+    this.suiKit =
+      instance?.suiKit ??
+      instance?.address?.cache._suiKit ??
+      new SuiKit(params);
+
+    if (instance?.address) {
+      this.address = instance.address;
+      this.cache = this.address.cache;
+      this.suiKit = this.address.cache._suiKit;
+    } else {
+      this.cache = new ScallopCache(this.suiKit, DEFAULT_CACHE_OPTIONS);
+      this.address =
+        instance?.address ??
+        new ScallopAddress(
+          {
+            id: params?.addressesId || ADDRESSES_ID,
+            network: params?.networkType,
+          },
+          {
+            cache: this.cache,
+          }
+        );
+    }
     this.isTestnet = params.networkType
       ? params.networkType === 'testnet'
       : false;
@@ -102,14 +107,10 @@ export class ScallopUtils {
    * @param address - ScallopAddress instance.
    */
   public async init(force: boolean = false, address?: ScallopAddress) {
-    if (force || !this._address.getAddresses() || !address?.getAddresses()) {
-      await this._address.read();
+    if (force || !this.address.getAddresses() || !address?.getAddresses()) {
+      await this.address.read();
     } else {
-      this._address = address;
-    }
-
-    if (!this._query.address.getAddresses()) {
-      await this._query.init(force, this._address);
+      this.address = address;
     }
   }
 
@@ -146,7 +147,7 @@ export class ScallopUtils {
   public parseCoinType(coinName: SupportCoins) {
     coinName = isMarketCoin(coinName) ? this.parseCoinName(coinName) : coinName;
     const coinPackageId =
-      this._address.get(`core.coins.${coinName}.id`) ||
+      this.address.get(`core.coins.${coinName}.id`) ||
       coinIds[coinName] ||
       undefined;
     if (!coinPackageId) {
@@ -154,20 +155,20 @@ export class ScallopUtils {
     }
     if (coinName === 'sui')
       return normalizeStructTag(`${coinPackageId}::sui::SUI`);
-    const wormHolePckageIds = [
-      this._address.get('core.coins.usdc.id') ?? wormholeCoinIds.usdc,
-      this._address.get('core.coins.usdt.id') ?? wormholeCoinIds.usdt,
-      this._address.get('core.coins.eth.id') ?? wormholeCoinIds.eth,
-      this._address.get('core.coins.btc.id') ?? wormholeCoinIds.btc,
-      this._address.get('core.coins.sol.id') ?? wormholeCoinIds.sol,
-      this._address.get('core.coins.apt.id') ?? wormholeCoinIds.apt,
+    const wormHolePackageIds = [
+      this.address.get('core.coins.usdc.id') ?? wormholeCoinIds.usdc,
+      this.address.get('core.coins.usdt.id') ?? wormholeCoinIds.usdt,
+      this.address.get('core.coins.eth.id') ?? wormholeCoinIds.eth,
+      this.address.get('core.coins.btc.id') ?? wormholeCoinIds.btc,
+      this.address.get('core.coins.sol.id') ?? wormholeCoinIds.sol,
+      this.address.get('core.coins.apt.id') ?? wormholeCoinIds.apt,
     ];
-    const voloPckageIds = [
-      this._address.get('core.coins.vsui.id') ?? voloCoinIds.vsui,
+    const voloPackageIds = [
+      this.address.get('core.coins.vsui.id') ?? voloCoinIds.vsui,
     ];
-    if (wormHolePckageIds.includes(coinPackageId)) {
+    if (wormHolePackageIds.includes(coinPackageId)) {
       return `${coinPackageId}::coin::COIN`;
-    } else if (voloPckageIds.includes(coinPackageId)) {
+    } else if (voloPackageIds.includes(coinPackageId)) {
       return `${coinPackageId}::cert::CERT`;
     } else {
       return `${coinPackageId}::${coinName}::${coinName.toUpperCase()}`;
@@ -222,7 +223,7 @@ export class ScallopUtils {
    * @returns sCoin treasury id
    */
   public getSCoinTreasury(sCoinName: SupportSCoin) {
-    return this._address.get(`scoin.coins.${sCoinName}.treasury`);
+    return this.address.get(`scoin.coins.${sCoinName}.treasury`);
   }
 
   /**
@@ -234,7 +235,7 @@ export class ScallopUtils {
    */
   public parseMarketCoinType(coinName: SupportCoins) {
     const protocolObjectId =
-      this._address.get('core.object') || PROTOCOL_OBJECT_ID;
+      this.address.get('core.object') || PROTOCOL_OBJECT_ID;
     const coinType = this.parseCoinType(coinName);
     return `${protocolObjectId}::reserve::MarketCoin<${coinType}>`;
   }
@@ -267,27 +268,27 @@ export class ScallopUtils {
 
     const wormHoleCoinTypeMap: Record<string, SupportAssetCoins> = {
       [`${
-        this._address.get('core.coins.usdc.id') ?? wormholeCoinIds.usdc
+        this.address.get('core.coins.usdc.id') ?? wormholeCoinIds.usdc
       }::coin::COIN`]: 'usdc',
       [`${
-        this._address.get('core.coins.usdt.id') ?? wormholeCoinIds.usdt
+        this.address.get('core.coins.usdt.id') ?? wormholeCoinIds.usdt
       }::coin::COIN`]: 'usdt',
       [`${
-        this._address.get('core.coins.eth.id') ?? wormholeCoinIds.eth
+        this.address.get('core.coins.eth.id') ?? wormholeCoinIds.eth
       }::coin::COIN`]: 'eth',
       [`${
-        this._address.get('core.coins.btc.id') ?? wormholeCoinIds.btc
+        this.address.get('core.coins.btc.id') ?? wormholeCoinIds.btc
       }::coin::COIN`]: 'btc',
       [`${
-        this._address.get('core.coins.sol.id') ?? wormholeCoinIds.sol
+        this.address.get('core.coins.sol.id') ?? wormholeCoinIds.sol
       }::coin::COIN`]: 'sol',
       [`${
-        this._address.get('core.coins.apt.id') ?? wormholeCoinIds.apt
+        this.address.get('core.coins.apt.id') ?? wormholeCoinIds.apt
       }::coin::COIN`]: 'apt',
     };
     const voloCoinTypeMap: Record<string, SupportAssetCoins> = {
       [`${
-        this._address.get('core.coins.vsui.id') ?? voloCoinIds.vsui
+        this.address.get('core.coins.vsui.id') ?? voloCoinIds.vsui
       }::cert::CERT`]: 'vsui',
     };
 
@@ -388,8 +389,8 @@ export class ScallopUtils {
     coinType: string = SUI_TYPE_ARG,
     ownerAddress?: string
   ) {
-    ownerAddress = ownerAddress || this._suiKit.currentAddress();
-    const coins = await this._suiKit.suiInteractor.selectCoins(
+    ownerAddress = ownerAddress || this.suiKit.currentAddress();
+    const coins = await this.suiKit.suiInteractor.selectCoins(
       ownerAddress,
       amount,
       coinType
@@ -412,14 +413,14 @@ export class ScallopUtils {
   ): Promise<void> {
     // merge to existing coins if exist
     try {
-      const existingSCoin = await this.selectCoins(
+      const existingCoins = await this.selectCoins(
         Number.MAX_SAFE_INTEGER,
         coinType,
         sender
       );
 
-      if (existingSCoin.length > 0) {
-        txBlock.mergeCoins(dest, existingSCoin);
+      if (existingCoins.length > 0) {
+        txBlock.mergeCoins(dest, existingCoins.slice(0, 500));
       }
     } catch (e) {
       // ignore
@@ -437,15 +438,15 @@ export class ScallopUtils {
    * @return Asset coin Names.
    */
   public async getObligationCoinNames(obligationId: SuiAddressArg) {
-    const obligation = await queryObligation(this._query, obligationId);
-    const collateralCoinTypes =
-      obligation?.collaterals.map((collateral) => {
-        return `0x${collateral.type.name}`;
-      }) ?? [];
-    const debtCoinTypes =
-      obligation?.debts.map((debt) => {
-        return `0x${debt.type.name}`;
-      }) ?? [];
+    const obligation = await queryObligation(this, obligationId);
+    if (!obligation) return undefined;
+
+    const collateralCoinTypes = obligation.collaterals.map((collateral) => {
+      return `0x${collateral.type.name}`;
+    });
+    const debtCoinTypes = obligation.debts.map((debt) => {
+      return `0x${debt.type.name}`;
+    });
     const obligationCoinTypes = [
       ...new Set([...collateralCoinTypes, ...debtCoinTypes]),
     ];
@@ -507,7 +508,7 @@ export class ScallopUtils {
       for (const endpoint of endpoints) {
         const priceIds = Array.from(failedRequests.values()).reduce(
           (acc, coinName) => {
-            const priceId = this._address.get(
+            const priceId = this.address.get(
               `core.coins.${coinName}.oracle.pyth.feed`
             );
             acc[coinName] = priceId;
@@ -520,14 +521,14 @@ export class ScallopUtils {
           Object.entries(priceIds).map(async ([coinName, priceId]) => {
             const pythConnection = new SuiPriceServiceConnection(endpoint);
             try {
-              const feed = await this._cache.queryClient.fetchQuery({
+              const feed = await this.address.cache.queryClient.fetchQuery({
                 queryKey: [priceId],
                 queryFn: async () => {
                   return await pythConnection.getLatestPriceFeeds([priceId]);
                 },
               });
               if (feed) {
-                const data = parseDataFromPythPriceFeed(feed[0], this._address);
+                const data = parseDataFromPythPriceFeed(feed[0], this.address);
                 this._priceMap.set(coinName as SupportAssetCoins, {
                   price: data.price,
                   publishTime: data.publishTime,
@@ -548,7 +549,7 @@ export class ScallopUtils {
       if (failedRequests.size > 0) {
         await Promise.allSettled(
           Array.from(failedRequests.values()).map(async (coinName) => {
-            const price = await this._query.getPriceFromPyth(coinName);
+            const price = await getPythPrice(this, coinName);
             this._priceMap.set(coinName, {
               price: price,
               publishTime: Date.now(),