+++ src/queries/sCoinQuery.ts
@@ -2,7 +2,7 @@ import { bcs } from '@mysten/sui.js/bcs';
 import assert from 'assert';
 import BigNumber from 'bignumber.js';
 import { SUPPORT_SCOIN } from 'src/constants';
-import { ScallopQuery } from 'src/models';
+import { ScallopQuery, ScallopUtils } from 'src/models';
 import { OptionalKeys, SupportSCoin, sCoinBalance } from 'src/types';
 
 /**
@@ -12,18 +12,22 @@ import { OptionalKeys, SupportSCoin, sCoinBalance } from 'src/types';
  * @returns `number`
  */
 export const getSCoinTotalSupply = async (
-  query: ScallopQuery,
+  {
+    utils,
+  }: {
+    utils: ScallopUtils;
+  },
   sCoinName: SupportSCoin
 ): Promise<sCoinBalance> => {
-  const sCoinPkgId = query.address.get('scoin.id');
+  const sCoinPkgId = utils.address.get('scoin.id');
   // get treasury
-  const args = [query.utils.getSCoinTreasury(sCoinName)];
+  const args = [utils.getSCoinTreasury(sCoinName)];
   const typeArgs = [
-    query.utils.parseSCoinType(sCoinName),
-    query.utils.parseUnderlyingSCoinType(sCoinName),
+    utils.parseSCoinType(sCoinName),
+    utils.parseUnderlyingSCoinType(sCoinName),
   ];
   const queryTarget = `${sCoinPkgId}::s_coin_converter::total_supply`;
-  const queryResults = await query.cache.queryInspectTxn({
+  const queryResults = await utils.cache.queryInspectTxn({
     queryTarget,
     args,
     typeArgs,
@@ -35,9 +39,7 @@ export const getSCoinTotalSupply = async (
     assert(type === 'u64', 'Result type is not u64');
 
     return BigNumber(bcs.de(type, value))
-      .shiftedBy(
-        query.utils.getCoinDecimal(query.utils.parseCoinName(sCoinName))
-      )
+      .shiftedBy(utils.getCoinDecimal(utils.parseCoinName(sCoinName)))
       .toNumber();
   }
 
@@ -53,17 +55,21 @@ export const getSCoinTotalSupply = async (
  * @return All owned sCoins amount.
  */
 export const getSCoinAmounts = async (
-  query: ScallopQuery,
+  {
+    utils,
+  }: {
+    utils: ScallopUtils;
+  },
   sCoinNames?: SupportSCoin[],
   ownerAddress?: string
 ) => {
   sCoinNames = sCoinNames || [...SUPPORT_SCOIN];
-  const owner = ownerAddress || query.suiKit.currentAddress();
+  const owner = ownerAddress || utils.suiKit.currentAddress();
   const sCoins = {} as OptionalKeys<Record<SupportSCoin, number>>;
 
   await Promise.allSettled(
     sCoinNames.map(async (sCoinName) => {
-      const sCoin = await getSCoinAmount(query, sCoinName, owner);
+      const sCoin = await getSCoinAmount({ utils }, sCoinName, owner);
       sCoins[sCoinName] = sCoin;
     })
   );
@@ -80,15 +86,87 @@ export const getSCoinAmounts = async (
  * @return Owned sCoin amount.
  */
 export const getSCoinAmount = async (
-  query: ScallopQuery,
+  {
+    utils,
+  }: {
+    utils: ScallopUtils;
+  },
   sCoinName: SupportSCoin,
   ownerAddress?: string
 ) => {
-  const owner = ownerAddress || query.suiKit.currentAddress();
-  const sCoinType = query.utils.parseSCoinType(sCoinName);
-  const amount = await query.cache.queryGetCoinBalance({
+  const owner = ownerAddress || utils.suiKit.currentAddress();
+  const sCoinType = utils.parseSCoinType(sCoinName);
+  const amount = await utils.cache.queryGetCoinBalance({
     owner,
     coinType: sCoinType,
   });
   return BigNumber(amount).toNumber();
+};
+
+const isSupportStakeCoins = (value: string): value is SupportSCoin => {
+  return SUPPORT_SCOIN.includes(value as SupportSCoin);
+};
+
+const checkAssetParams = (fromSCoin: SupportSCoin, toSCoin: SupportSCoin) => {
+  if (fromSCoin === toSCoin)
+    throw new Error('fromAsset and toAsset must be different');
+
+  if (!isSupportStakeCoins(fromSCoin))
+    throw new Error('fromAsset is not supported');
+
+  if (!isSupportStakeCoins(toSCoin)) {
+    throw new Error('toAsset is not supported');
+  }
+};
+
+/* ==================== Get Swap Rate ==================== */
+
+/**
+ * Get swap rate from sCoin A to sCoin B.
+ * @param fromSCoin
+ * @param toSCoin
+ * @param underlyingCoinPrice - The price of the underlying coin. For example, if fromSCoin is sSUI and toSCoin is sUSDC, then underlyingCoinPrice represents the price of 1 SUI in USDC.
+ * @returns number
+ */
+export const getSCoinSwapRate = async (
+  query: ScallopQuery,
+  fromSCoin: SupportSCoin,
+  toSCoin: SupportSCoin,
+  underlyingCoinPrice?: number
+) => {
+  checkAssetParams(fromSCoin, toSCoin);
+  const fromCoinName = query.utils.parseCoinName(fromSCoin);
+  const toCoinName = query.utils.parseCoinName(toSCoin);
+
+  // Get lending data for both sCoin A and sCoin B
+  const marketPools = await Promise.all([
+    query.getMarketPool(fromCoinName, false),
+    query.getMarketPool(toCoinName, false),
+  ]);
+  if (marketPools.some((pool) => !pool))
+    throw new Error('Failed to fetch the lendings data');
+
+  if (marketPools.some((pool) => pool?.conversionRate === 0)) {
+    throw new Error('Conversion rate cannot be zero');
+  }
+
+  const ScoinAToARate = marketPools[0]!.conversionRate;
+  const BtoSCoinBRate = 1 / marketPools[1]!.conversionRate;
+
+  const calcAtoBRate = async () => {
+    const prices = await query.utils.getCoinPrices([fromCoinName, toCoinName]);
+    if (!prices[fromCoinName] || !prices[toCoinName]) {
+      throw new Error('Failed to fetch the coin prices');
+    }
+    if (prices[toCoinName] === 0) {
+      throw new Error('Price of toCoin cannot be zero');
+    }
+    return prices[fromCoinName]! / prices[toCoinName]!;
+  };
+
+  const AtoBRate = underlyingCoinPrice ?? (await calcAtoBRate());
+  return BigNumber(ScoinAToARate)
+    .multipliedBy(AtoBRate)
+    .multipliedBy(BtoSCoinBRate)
+    .toNumber();
 };